src/Pyrope.AISidecar/requirements.txt
grpcio
grpcio-tools
protobuf
scikit-learn
numpy
requests
google-generativeai
